# packages/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Vexel Health Platform API
  version: 1.0.0
  description: |
    Official API contract for Vexel Health Platform.

    Production tenancy is resolved by request Host (TenantDomain mapping).
    Clients do NOT choose tenant via header in production.
    The `x-tenant-id` header is allowed only in local DEV when `TENANCY_DEV_HEADER_ENABLED=1`.
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /encounters:
    post:
      summary: Create encounter
      operationId: createEncounter
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEncounterDto'
      responses:
        '201':
          description: Encounter created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Encounter'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/AuthError'
        '409':
          $ref: '#/components/responses/DomainError'
        '500':
          $ref: '#/components/responses/UnexpectedError'
    get:
      summary: Search encounters
      operationId: listEncounters
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: patientId
          in: query
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of encounters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEncountersResponse'
        '500':
          $ref: '#/components/responses/UnexpectedError'
...
  /lab/tests:
    post:
      summary: Create LAB test definition
      operationId: createLabTest
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTestRequest'
      responses:
        '201':
          description: LAB test created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabTestDefinition'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/DomainError'
        '500':
          $ref: '#/components/responses/UnexpectedError'
    get:
      summary: List LAB test definitions
      operationId: listLabTests
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      responses:
        '200':
          description: LAB test list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListLabTestsResponse'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /lab/tests/{testId}:
    get:
      summary: Get LAB test definition by id
      operationId: getLabTestById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/LabTestIdPathParam'
      responses:
        '200':
          description: LAB test definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabTestDefinition'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /lab/tests/{testId}/parameters:
    post:
      summary: Add parameter to LAB test definition
      operationId: addLabTestParameter
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/LabTestIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddParameterRequest'
      responses:
        '201':
          description: LAB parameter created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabTestParameter'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/AuthError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/UnexpectedError'
    get:
      summary: List parameters for LAB test definition
      operationId: listLabTestParameters
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/LabTestIdPathParam'
      responses:
        '200':
          description: LAB parameters list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListLabTestParametersResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/UnexpectedError'
...
  /encounters/{id}/lab-tests:
    get:
      summary: List ordered LAB tests with parameters and results for encounter
      operationId: listEncounterLabTests
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/IdPathParam'
      responses:
        '200':
          description: Ordered LAB tests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEncounterLabTestsResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/DomainError'
        '500':
          $ref: '#/components/responses/UnexpectedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantIdHeader:
      name: x-tenant-id
      in: header
      description: DEV ONLY. Ignored in production. For local testing when hostname is localhost.
      required: false
      schema:
        type: string
        format: uuid
    IdPathParam:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    LabTestIdPathParam:
      name: testId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Structured validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorEnvelope'
    DomainError:
      description: Structured domain rule error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DomainErrorEnvelope'
    AuthError:
      description: Authentication or authorization error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthErrorEnvelope'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFoundErrorEnvelope'
    UnexpectedError:
      description: Unexpected server-side failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnexpectedErrorEnvelope'
  schemas:
...
    CreateEncounterDto:
      type: object
      required:
        - patientId
        - type
      properties:
        patientId:
          type: string
          format: uuid
        type:
          type: string
          enum: [LAB, RAD, OPD, BB, IPD]
        startedAt:
          type: string
          format: date-time
    Encounter:
      allOf:
        - $ref: '#/components/schemas/CreateEncounterDto'
        - type: object
          required:
            - id
            - tenantId
            - encounterCode
            - status
            - createdAt
            - startedAt
          properties:
            id:
              type: string
            tenantId:
              type: string
            encounterCode:
              type: string
              example: OPD-2026-000001
            status:
              type: string
              enum: [CREATED, PREP, IN_PROGRESS, FINALIZED, DOCUMENTED]
            endedAt:
              type: string
              format: date-time
            createdAt:
              type: string
              format: date-time
    ListEncountersResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Encounter'
        total:
          type: integer
    LabTestDefinition:
      type: object
      required:
        - id
        - code
        - name
        - department
        - active
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        department:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LabTestParameter:
      type: object
      required:
        - id
        - testId
        - name
        - displayOrder
        - active
      properties:
        id:
          type: string
          format: uuid
        testId:
          type: string
          format: uuid
        name:
          type: string
        unit:
          type: string
          nullable: true
        refLow:
          type: number
          nullable: true
        refHigh:
          type: number
          nullable: true
        refText:
          type: string
          nullable: true
        displayOrder:
          type: integer
        active:
          type: boolean
    LabOrderItemStatus:
      type: string
      enum: [ORDERED, RESULTS_ENTERED, VERIFIED]
    LabResultFlag:
      type: string
      enum: [LOW, HIGH, NORMAL, ABNORMAL, UNKNOWN]
    LabOrderItem:
      type: object
      required:
        - id
        - encounterId
        - testId
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        encounterId:
          type: string
          format: uuid
        testId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/LabOrderItemStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LabResultItem:
      type: object
      required:
        - id
        - orderItemId
        - parameterId
        - value
        - flag
      properties:
        id:
          type: string
          format: uuid
        orderItemId:
          type: string
          format: uuid
        parameterId:
          type: string
          format: uuid
        value:
          type: string
        valueNumeric:
          type: number
          nullable: true
        flag:
          $ref: '#/components/schemas/LabResultFlag'
        enteredBy:
          type: string
          nullable: true
        enteredAt:
          type: string
          format: date-time
          nullable: true
        verifiedBy:
          type: string
          nullable: true
        verifiedAt:
          type: string
          format: date-time
          nullable: true
    LabOrderedTest:
      type: object
      required:
        - orderItem
        - test
        - parameters
        - results
      properties:
        orderItem:
          $ref: '#/components/schemas/LabOrderItem'
        test:
          $ref: '#/components/schemas/LabTestDefinition'
        parameters:
          type: array
          items:
            $ref: '#/components/schemas/LabTestParameter'
        results:
          type: array
          items:
            $ref: '#/components/schemas/LabResultItem'
    ListLabTestsResponse:
      type: object
      required:
        - data
        - total
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/LabTestDefinition'
        total:
          type: integer
    ListLabTestParametersResponse:
      type: object
      required:
        - data
        - total
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/LabTestParameter'
        total:
          type: integer
    ListEncounterLabTestsResponse:
      type: object
      required:
        - data
        - total
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/LabOrderedTest'
        total:
          type: integer
    CreateTestRequest:
      type: object
      additionalProperties: false
      required:
        - code
        - name
        - department
      properties:
        code:
          type: string
        name:
          type: string
        department:
          type: string
        active:
          type: boolean
    AddParameterRequest:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
        unit:
          type: string
        refLow:
          type: number
        refHigh:
          type: number
        refText:
          type: string
        displayOrder:
          type: integer
        active:
          type: boolean
...
    ValidationErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            type:
              type: string
              enum: [validation_error]
            fields:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
          required:
            - type
            - fields
      required:
        - error
    DomainErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            type:
              type: string
              enum: [domain_error]
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
          required:
            - type
            - code
            - message
      required:
        - error
    AuthErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            type:
              type: string
              enum: [auth_error]
            message:
              type: string
          required:
            - type
            - message
      required:
        - error
    NotFoundErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            type:
              type: string
              enum: [not_found]
            message:
              type: string
          required:
            - type
            - message
      required:
        - error
    UnexpectedErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            type:
              type: string
              enum: [unexpected_error]
            message:
              type: string
            correlationId:
              type: string
          required:
            - type
            - message
      required:
        - error